Move chord structs to chords.h file
